[File,Sugar] Better javascript data files integration. Refs #29
This commit not only removes the dependency to jsinterpreter, but also provide
a native binding to the require function of the NodeJS module scope.

In short: it is now possible to use a NodeJS compatible javascript file as a data
source for doc-e-mate documents. The only requirement is to export the data with
a 'module.exports = data;' instruction at the end of the javascript.