Make tests pass.
This commit make the acceptance tests pass.
But it implies a minor regression in the expected output in the case of the help
commands. There is currently a bug which make the output of those command be
truncated at the end. It misses the two following lines:
```
Please report bugs at https://github.com/enspirit/dem-rescript

```
This should be fixed once migrated from bs-cmdliner to an equivalent as
commander.js or yargs.